+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # History
 
-## Unreleased Changes
+## 2.1.0 (2018-08-21)
 
 ### Added